requirements.txt
asgiref==3.3.1
Django==3.1.5
django-location-field==2.1.0
Pillow==8.1.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
gunicorn-20.0.4 
whitenoise-5.2.0
